+++ config/theme.go
@@ -60,7 +60,7 @@ var theme = map[string]tcell.Color{
 	"link_number":       tcell.ColorSilver,
 	"regular_text":      tcell.ColorWhite,
 	"quote_text":        tcell.ColorWhite,
-	"preformatted_text": tcell.ColorGrey,
+	"preformatted_text": tcell.Color229, // xterm:Wheat1, #ffffaf
 	"list_text":         tcell.ColorWhite,
 }